🔨 Rename user to users
