+++ src/users/entities/user.interface.ts
