+++ src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { UsersController } from './users.controller';
 import { DynamooseModule } from 'nestjs-dynamoose';
-import { UserSchema } from './user.schema';
+import { UserSchema } from './entities/user.schema';
 
 @Module({
   imports: [
